권한처리 코드 추가

+++ src/main/java/kr/ac/deu/computer_engineering/Absenteeism/Management/route/user/UserController.java
@@ -3,18 +3,23 @@ package kr.ac.deu.computer_engineering.Absenteeism.Management.route.user;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.User.User;
-import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.User.dto.CreateUserDto;
-import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.User.dto.UpdateUserDto;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.User.dto.UserDto;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.User.dto.UserDto;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.service.user.UserService;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.utils.RoleValidate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Slf4j
 @RestController
@@ -28,9 +33,16 @@ public class UserController {
             summary = "직원 목록 조회",
             description = "직원 목록 조회")
     @GetMapping("")
-    public ResponseEntity<?> getItemList(@RequestParam(required = false) String q) {
-        List<User> userList = userService.getList(q);
-        return new ResponseEntity<>(userList, HttpStatus.OK);
+    public ResponseEntity<?> getItemList(
+            HttpServletRequest request,
+            @RequestParam(required = false) String q) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            List<User> userList = userService.getList(q);
+            return new ResponseEntity<>(userList, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "직원")
@@ -38,9 +50,17 @@ public class UserController {
             summary = "직원 Id로 상세 조회",
             description = "직원 상세 조회")
     @GetMapping("/{userId}")
-    public ResponseEntity<?> getItemById(@PathVariable Long userId) throws Exception {
-        User user = userService.getUserById(userId);
-        return new ResponseEntity<>(user, HttpStatus.OK);
+    public ResponseEntity<?> getItemById(
+            HttpServletRequest request,
+            @PathVariable Long userId) throws Exception {
+        HttpSession session = request.getSession();
+        Long sUserId = RoleValidate.getUserId(session);
+        if (Objects.equals(sUserId, userId) || RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            User user = userService.getUserById(userId);
+            return new ResponseEntity<>(user, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "직원")
@@ -48,9 +68,17 @@ public class UserController {
             summary = "직원 정보 생성",
             description = "직원 정보 생성")
     @PostMapping("")
-    public ResponseEntity<?> createItem(@RequestBody CreateUserDto dto) throws Exception {
-        userService.createUser(dto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public ResponseEntity<?> createItem(
+            HttpServletRequest request,
+            @Valid @RequestBody UserDto dto,
+            BindingResult bindingResult) throws Exception {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            userService.createUser(dto);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "직원")
@@ -59,10 +87,18 @@ public class UserController {
             description = "직원 정보 수정")
     @PutMapping("/{userId}")
     public ResponseEntity<?> updateItemById(
-            @RequestBody UpdateUserDto dto,
+            HttpServletRequest request,
+            @Valid @RequestBody UserDto dto,
+            BindingResult bindingResult,
             @PathVariable Long userId) throws Exception {
-        userService.updateUser(userId, dto);
-        return new ResponseEntity<>(HttpStatus.OK);
+        HttpSession session = request.getSession();
+        Long sUserId = RoleValidate.getUserId(session);
+        if (Objects.equals(sUserId, userId) || RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            userService.updateUser(userId, dto);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "직원")
@@ -70,8 +106,15 @@ public class UserController {
             summary = "직원 Id로 직원 정보 삭제",
             description = "직원 정보 삭제")
     @DeleteMapping("/{userId}")
-    public ResponseEntity<?> deleteItemById(@PathVariable Long userId) throws Exception {
-        userService.deleteUser(userId);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<?> deleteItemById(
+            HttpServletRequest request,
+            @PathVariable Long userId) throws Exception {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            userService.deleteUser(userId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 }
